Enable OpenPower SBE Vital Attention handler
If PRD notices that a sbe vital attention is set (TP_LFIR 26) then
it will call hostboot code to attempt to recover the SBE. If this
occurs during IPL time then hostboot will not be able to recover
the SBE and we will deconfigure the processor. If this occurs
during runtime HBRT will attempt to run the retry_handler. This will
result in us calling hreset on the SBE that failed. If we were able
to recover the SBE then no error will be returned. If we are unable
to recover the SBE then we will return an error with a deconfig
record.

Change-Id: I3da6ec932ef8e59f7b2a184621a47e88d465e0c5
RTC: 167191
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/56821
CI-Ready: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>