Cleanup the fix for "zero length dump on single node systems".
The attribute, IS_MASTER_DRAWER, was being set in the function
host_sys_fab_iovalid_processing() (istep 18.9), unfortunately that
function is not called on single node systems.

Had to reverse the logic on the previous IS_MASTER_DRAWER attribute,
and make it an IS_SLAVE_DRAWER attribute.  The reason is that there
is a limitation (bug?) which requires hb-only, volatile attributes
to have a default value of zero.

Change-Id: Ib27349f18b2738b405649723cdcb11e2ec919991
CQ:SW423016
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/56883
CI-Ready: Richard Ward <rward15@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Richard J. Knight <rjknight@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>