Improve linking of PLIDs for sbe_retry_handler
Previously if a PLID was passed to the ctor of the sbe_retry_handler
we would link all errors logs created during the process of recovering
the SBE with this PLID. But if no PLID was passed then we would not
link the logs. This commit changes it so if no PLID is passed to the
ctor then the first log created in the recovery process will become
the PLID that all logs after will be set to.

Change-Id: I93ef3a48b4cc1d7df3237d7ba3dfefba21d5fb6b
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/56885
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>