Clear perv scrach register prior to triggering HRESET
Before we initiate an HRESET we need to make sure that the perv
scratch register that the SBE looks at during boot are all zeros.
When these registers are zeros the SBE will ignore them and pull
values from its own image. We need to do this because Hostboot will
repurpose these registers after the SBE boots the first time so the
data in the registers is no longer valid. It is okay for the SBE to
pull the values from it's own image because during the IPL hostboot
customized the SBE image with the correct values.

Change-Id: I8b434d04cde3c384e35a3089a349a1d121b6b1dc
RTC: 180242
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/56959
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
CI-Ready: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>