Secure Boot: Inhibit preverifying WOFDATA reserved memory area
Hostboot reserved memory areas are large, and there is enough content in
certain cases such that data bleeds into the start of the PHyp image.  This
change stops preverifying one of the largets partitions, WOFDATA, so that
the reserved memory areas will stay well within bounds for the time being.

Change-Id: Ia099bc3bc27ba7000c2e5a1420a1e8ac72537ba6
CQ: SW421678
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57031
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>