Check for pending OCC messages on HBRT start in OpenPOWER
The OCC will set an interrupt for the Host to look for a pending
message/action.  If opal-prd is not running at the time that this
interrupt happens, HTMGT will never see the message.  Therefore
we need to have HTMGT explicitly check for anything pending any
time HBRT starts up.

Change-Id: Iaae075ae6828a9de5a7bd5afe55c541b0076f2d2
RTC: 187912
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57175
Reviewed-by: Christopher J. Cain <cjcain@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>