Ensure that we only attempt to recover from vital attn on BMC systems
On an FSP system PRD should never get triggered on an SBE vital attn.
However if for any reason PRD attempts to handle a vital attn on an
FSP system HB should not actually perform any recovery and we should
just return an error back to PRD signaling that we could not recover.
PRD should stop mask off the error after the handler returns with
an error saying we could not recover.

Change-Id: If1666bd028da9a94bed34d79da34f6264beb0fbf
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57176
CI-Ready: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>