Add script to generate istep wait timer attribute override
This commit introduces a simple script that will generate an attr
override file that the user can easily load into /tmp/ on an FSP
in order to add a delay after a given major/minor istep combination.
This script can be used by both hostboot developers or by the test
team.

Change-Id: If9c15d5c2988eba28e026193e26d616c68ebc881
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57303
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Brian E. Bakke <bbakke@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>