Drain message queue prior to setting iv_disabled in mailboxsp
Set iv_disabled after we have resumed the mbox interrupts so that the
message queue can drain. If iv_disabled is true, resume() will not do
anything and the mbox msg queue will not drain. This can cause us to
hang in the shutdown path because we will never reach the quiesced
state that we need to be in before initiating handleShutdown()

Change-Id: Ibf1dc1d59a496d3b33fc9fca5080349d945c0169
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57439
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Brian E. Bakke <bbakke@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>