Rediscover I2C Targets after Host I2C Reset
It's possible that some targets that use I2C to detect if they are
present  might not be present after a system re-IPLs and performs a
FSI I2C Reset.  Therefore, later in the IPL, after a Host I2C Reset
sequence an attempt to rediscover any of these targets is made.
If any new targets are found, the system will re-IPL.  Currently
only supporting looking for DIMMs now on OpenPower systems with
securemode enabled.

Change-Id: I010135231f2f74869529a3dbc3344413b6d19dc9
RTC:178973
Backport:release-op920
CQ:SW427365
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/58125
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>