Move VPD cache invalidation in mfgmode after we save it for HBRT
Inside call_host_runtime_setup() there is a call to
VPD::goldenCacheInvalidate().  This function invalidates our VPD
cache in PNOR in mfg-mode in order to allow manufacturing to modify
serial numbers at runtime and have those settings take effect on
the next boot.  This has been in place since early P8 OpenPOWER.

In that same function is a call to RUNTIME::populate_hbRuntimeData().
This is where we copy the Hostboot data out into mainstore for HBRT
to use later on.  One of those pieces of data is the VPD cache.

In January a fix went in for a different issue that moved where
populate_hbRuntimeData() was run, putting it *after* we clear it in
mfgmode.  Obviously that isn't going to work so well...

This commit moves the invalidation later in the function so that it
is after we save off the cache again.

Change-Id: Idb482b583ecdb7bb8abcb617bae4ae786c110fea
CQ: SW424124
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57469
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>