Adds workaround for missing attributes from VPD
The current rendition of the Moby Dick board does not have
ATTR_CEN_VPD_ISDIMMTOC4DQ or ATTR_CEN_VPD_ISDIMMTOC4DQS.
As updating all of the current boards out there will take a long time,
a code workaround to handle the failing case.  We log the failing error,
print an error asking the user to contact a memory representative,
and fill the array with all zero values.

Change-Id: Id8f6de14a4ebd2e85422e7c82381636062b095d3
RTC:189874
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57555
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: ANDRE A. MARIN <aamarin@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57567
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>