Remove Ordinal ID as an Attribute to Match
During the preparation for doing an HBRT concurrent code update, targets in
the current reserved memory data are matched with targets in the new LID
structure.  One of the attributes used has been ordinal ID, but this is not
always set in the LID structure, thus preventing matches to be found.  It is
being removed as a condition for finding a match.

Change-Id: I6b0bea6a001a9ca085fb5a41da714780543655e6
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57583
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Brian E. Bakke <bbakke@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>