Improve resource recovery path to handle memory plugging rules
Resource recovery is a RAS feature wherein we do not apply certain gard records
if those records would result in a boot fail due to a lack of hardware.  Change
the logic from applying speculative deconfiguration one by one to applying all
predictive gard records and then removing those records if minimal hardware is
not available.  Also add BLOCK_SPEC_DECONFIG attribute to flag when speculative
deconfiguration is not allowed.  This flag is set if minimum hardware is not
available and is cleared after a hardware change.

Change-Id: Ia065de3a44ab29fbf33ad4ce98bc42ea5144463f
RTC: 191414
CQ: SW424137
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/58041
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>