wrap_test: adding attrs to swizzle abus links
There is a fabric limitation to only enable 2 ABUS links at
a time. This commit adds attribute that will be used
by MST to swizzle between the sets of links to enable.
When the value of the attribute matches between system
target and the ABUS link, then the link will be
configured. Otherwise, we will deconfigure the links.

Change-Id: I018b318934b3f52c2e73724cd0e9645e5a379251
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57679
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>