Update targeting init during RT to set PEER_TARGET on OBUS targets
During IPL time hostboot does not know about OBUS targets on other
processors. Because of this we cannot set the PEER_TARGET attribute
on the OBUS targets because we don't know about their peers. During
HBRT however, we do know this information. In this commit support is
added to look at the attribute PEER_PATH for targets whose PEER_TARGET
attribute is null. This will allow use to target and OBUS's peer
during HBRT.

Change-Id: I249bc8c60297eedb7282ecc8dc717973e81e5cc3
CQ: SW425530
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57782
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Benjamin J. Weisenbeck <bweisenb@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>