genNodeMrwXml.pl: added --config option
Without the config option, the script was computing the MRW name
and output file names based on system name and build type. But,
for fleetwood's 2N and MST config, we were having to do unnecessary
copies to and from 2N/MST version of MRW to non 2N/MST version.
This causes issues when multiple compile jobs are running in
parallel.

Change-Id: Ib4e2d10b05f182379fd0be1cacbf1a7b9efba642
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57820
Reviewed-by: Elizabeth K. Liner <eliner@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>