Primary TPM Required Policy
Change the "TPM Required" policy to "Primary TPM Required," wherein
only the presence and functionality of the primary TPM (attached to
the acting master proc) will be considered when evaluating the
TPM required condition. The presence and functionality of the backup
TPM will not play a role in this policy.

Change-Id: Id968123d0fc399c531da7429fdb8efabfa66c53c
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57843
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>