Create HWP to save/restore OBUS fir masks
During the IPL in a multi-node system Hostboot is unable to detect other
nodes. Because of this Hostboot is unable to set the PEER_TARGET attr
for OBUS targets as their PEERs are on different nodes. Once the system
reaching runtime HBRT will have the information to fill in PEER_TARGET
attributes on the OBUS targets. During the period while Hostboot doesn't
know the OBUS PEER_TARGETs we must mask off additional FIRs so that PRD
doesn't attempt to access the an OBUS's peer during a fail. This HWP will
provide a way to save off the FIR masks during IPL time, and will apply
additional masks required for the rest of the IPL. In addition this HWP
will provide a way to reapply the original masks after HBRT has filled
in the information for all OBUSes peer targets.

CQ: SW425530
Change-Id: I1553b698aec050a94344592952fe7fda73871973
Original-Change-Id: I45a3d23023f0021d7dbf5c6726075b270e2e1418
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57853
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Reviewed-by: Christopher W. Steffen <cwsteffen@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>