Better debug for weird PIR issues
Ran into a case where Hostboot was getting booted with the wrong
fabricid for the node it was on.  In the process I realized it
was hard to tell what was going on.

Added printk output to show the PIR of the boot core.

Modified how force-hrmor option is consumed in ecmd-debug-framework
so that you can force unnatural lookups.

Change-Id: I77fc614889aa3cedcc8ccd39cac9ebc36f239b56
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/57910
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>