Deconfig record can't get added to vitalAttn elog on FSP
On FSP system we should never try to call vitalAttnHandler because
the fir bits that trigger this code should be masked. However
we have seen issues where for whatever reason the firs get unmasked.
If we happen to fall into this code on a FSP system we want to make
sure that no deconfigure records get added to the error logs that
get generated. If deconfig records get added it can collide with
the flow that hwsv is doing to attempt to recover the SBE. Also
in this commit we removed the TODO for disabling the OCC if the SBE
dies. We have decided that it is no longer necessary to disable
the OCC if the SBE dies. See story for more information.

Change-Id: Ib75dc34a122a37853047af67ec8858dbca11e4a2
CQ: SW427127
RTC: 180244
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/58226
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Brian E. Bakke <bbakke@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>