Check if Console Library is Loaded Before Printing to Console
When a failure occurs early in IPL (for instance, due to
corrupt secure headers), the doShutdown path attempts to
print a trace to the console, which, if the console has
not been initialized yet, locks up the shutdown path, and
we don't shut down gracefully. This change adds a check
to make sure the console library has been loaded prior to
flushing the trace in doShutdown path.

Change-Id: I74942476de0a8c3f4cba41b92b05369687441f68
CQ: SW426722
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/58346
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>