istepdispatcher: When stopping the watchdog, clear DONT_STOP
We intend to actually stop the watchdog. Don't set the DONT_STOP bit
in the set request as this keeps the watchdog ticking and will generate
a log entry when it expires. Instead allow the set command stop it
as defined in the IPMI spec by not sending DONT_STOP.

Signed-off-by: William A. Kennington III <wak@google.com>