Add option to masterProcChipTargetHandle() to return functional chip
This commit adds an option to the two masterProcChipTargetHandle()
functions to only look for and return a functional master proc chip
as necessary with the issue in Defect SW424528.  The default behavior
of these functions - where the functionality is not checked - remains
the same.  Once use of these functions has also been updated to use
this new option.

Change-Id: I37049d2cb9299a9404b57d85031a364bdb257c82
CQ:SW424528
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/58920
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>