Update getChildTargetsForCDG to use CEN PORT/DIMM
The port and dimm sockets come off of the centaur on cumulus.
Update this function to look at the correct attributes.

These attribute do not exist on Cumulus so the default will
be to just return all dimms under the input target.

This code change affected the test case so an update was required
there as well (i.e. the fix for this defect).

Change-Id: I102c40acc124100e90dd832aeed94716c0f18811
CQ:SW428864
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/59073
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>