Deprecate legacy ATTR_MBA_PORT/ATTR_MBA_DIMM
There are currently two competing sets of attributes related to
the DIMM position.
- Legacy (P8) MBA_PORT/MBA_DIMM
- New (P9) CEN_MBA_PORT/CEN_MBA_DIMM

These two sets of attributes are not currently being populated
with the same data in all of the MRW xml files.  The P9 HWPs are
using the CEN_ versions but there is a lot of code on the FSP
that is still using the old versions.  Additionally, the Nimbus
MRW XMLs are using MBA_PORT to represent the dimm position
behind the MCA (even though there is only a single port on the
MCA and MBA is erroneous).

This commit attempts to consolidate things into a new set of
more generic (and hopefully more descriptive) attributes to use
for all system types.  The MRW parsing logic has been modified
to consume whatever data it can to populate all 3 sets of data
with the same values.  Eventually we will deprecate the 2
previous attribute sets once we've cleaned all the existing files
up.

The new attributes are:
- MEM_PORT
- POS_ON_MEM_PORT

CQ: SW430317
Change-Id: If63929d78dcaacc1120709a659d6094ce576918d
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/60492
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>