ipmiwatchdog: Handle uninitialized errors during reset
The IPMI spec allows for the IPMI watchdog to return an error if the
watchdog was never initialized. This can happen if the daemon restarts
or if the initial setWatchDog command fails for some reason. Instead of
allowing the watchdog to remain uninitialized and disabled, use the
values of the previously issued set command to re-initialize the
watchdog. This is only attempted once per resetWatchDog call, but
periodic calls to the resetWatchDog should eventually re-initialize it.

Signed-off-by: William A. Kennington III <wak@google.com>