ipmiwatchdog: Handle uninitialized errors during reset
The IPMI spec allows for the IPMI watchdog to return an error if the
watchdog was never initialized. This can happen if the daemon restarts
or if the initial setWatchDog command fails for some reason. Instead of
allowing the watchdog to remain uninitialized and disabled, use the
values of the previously issued set command to re-initialize the
watchdog. This is only attempted once per resetWatchDog call, but
periodic calls to the resetWatchDog should eventually re-initialize it.

Resolves #139

Signed-off-by: William A. Kennington III <wak@google.com>
Change-Id: I7b777e2f3b75d5ce38888018c1b74c6ce12ffc19
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/59163
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>