Add clock callout enums for specific clock sources
PRD running up in the host needs to be able to point at a specific
clock source when detecting some errors.  Since the Hostboot
targeting model doesn't include the clock targets, we will use
the existing clock callout pathway to do this.  This change adds
a few new constants that specific which clock has the issue
instead of using the 'currently active' clock that we have now.

Change-Id: I6441ba3d1a082134cf2ada37cefca4cba1781924
RTC: 190666
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/59133
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Manish K. Chowdhary <manichow@in.ibm.com>
Reviewed-by: Benjamin J. Weisenbeck <bweisenb@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>