Fix indirect scoms at runtime under Opal
A recent change centralized some error checking code and in the
process broke indirect scoms under Opal.  There is a difference
of behavior between Opal and Phyp for these kinds of addresses.
In Phyp, HBRT does all of the translation into physical addresses.
In Opal, we send the complete 64-bit address to Opal for them
to handle it, therefore we need to allow it through our scom
code.

Resolves boston-openpower/#1293

Change-Id: I652106ed36dbfa3d2abcffd5fbf0d93f51bb631e
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/59156
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>