Defer handling attentions on non-master proc until after SMP is up
The current code enables PRD/Attention Handling on the slave
processors as soon as we get the SBE started (clocks running).
This is a problem because most of the FIR regs are blacklisted
such that we cannot write them until after the SMP is up and
we are using xscom.  This restriction will lead to an infinite
loop in PRD if we get a recoverable attention on the slave proc.

The fix is to defer all of the attention handling on the slave
procs until after the SMP is up.  We will move the enabled from
istep 8.5 host_attnlisten_proc to the end of istep 10.1
proc_build_smp.  Note that this makes 8.5 a complete NOOP.

Change-Id: I1b6542efe2cd14717d7fa55d01327121027c6862
CQ: SW429438
Backport: yes
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/59268
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane C. Shelley <zshelle@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>