Add procedure for istep 18's Node Communication ABUS Exchange
This commit creates a new interface that will be called in istep 18
to facilitate the exchanging of Link Mailboxes messages between nodes
acros the ABUS interface.  It also does the basic procedure without
all of the final FFDC and RAS implementation that will come in
subsequent commits.  All error logs are deleted to not kill any
IPLs until full functionality is tested.

Change-Id: I7391deaa37e49629cd01cfad212b64fecb9529bd
RTC:184518
CMVC-Prereq:1058895
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/60793
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>